Link List Assignment Code
Change all the variable names and delete all the comments. Comments are there for you to easily understand the code and get the idea. Here I've created a Link class and it's not necessary, you can write all the methods inside the LinkList class and call them in main method of LinkListApp class.